refactor(exec_command): split exec_command_impl and validation.rs for #1221
