+++ crates/aptu-coder/src/lib.rs
@@ -30,6 +30,7 @@ pub(crate) mod logging;
 pub(crate) mod metrics;
 pub(crate) mod otel;
 pub(crate) mod shell;
+pub(crate) mod shell_write;
 pub(crate) mod tools;
 pub(crate) mod validation;
