+++ crates/aptu-coder/src/validation.rs
@@ -6,440 +6,6 @@ use rmcp::model::ErrorData;
 
 use crate::tools::common::error_meta;
 
-/// Scans a shell command string for unclosed heredocs and file-write heredoc
-/// patterns before any process is spawned.
-///
-/// Phase 1 pre-scan: walks the command byte-by-byte with quote tracking.  When a
-/// `<<` token is found outside any quoted region, it scans backward to check for
-/// file-write patterns (cat/tee/redirect + `<<`) and returns an error immediately
-/// if one is detected.
-///
-/// Phase 2 main scan: continues the existing matching-closer scan (unchanged logic)
-/// with quote state reset to initial values between phases.
-///
-/// Returns `Ok(())` if no file-write patterns or unclosed heredocs are found,
-/// or `Err(ErrorData)` with `INVALID_PARAMS` otherwise.
-///
-/// This function does NOT spawn any process; it is a pure string scan.
-pub(crate) fn validate_heredocs(command: &str) -> Result<(), ErrorData> {
-    let bytes = command.as_bytes();
-    let len = bytes.len();
-
-    // Phase 1: pre-scan for heredoc file-write patterns (redirect + <<)
-    //
-    // Supported patterns (all rejected):
-    //   - cat > file << EOF
-    //   - cat >> file << EOF
-    //   - tee file << EOF, tee -a file << EOF
-    //   - tee > file << EOF, tee >> file << EOF
-    //   - printf 'content' > file << EOF
-    //   - dd of=file << EOF
-    //   - install file << EOF
-    //   - cp /dev/stdin file << EOF
-    //   - mv /dev/stdin file << EOF
-    //   - $VAR > file << EOF (variable-expanded command name)
-    //   - > file << EOF, >> file << EOF (bare redirect)
-    //   - (cat > file << EOF) (subshell grouping)
-    //   - cat > >(proc) << EOF (process substitution in file position)
-    //   - cat > $(cmd) << EOF (command substitution in file position)
-    //
-    // Unsupported (NOT rejected -- heuristic misses complex nesting):
-    //   - cat > "${var}" << EOF (variable expansion in file path)
-    //   - exec 3<>file; cat >&3 << EOF (arbitrary fd redirects)
-    //   - deeply nested subshells with multiple heredoc levels
-    {
-        let mut i = 0;
-        let mut in_single_quote = false;
-        let mut in_double_quote = false;
-
-        while i < len {
-            let ch = bytes[i] as char;
-
-            if ch == '\'' && !in_double_quote {
-                in_single_quote = !in_single_quote;
-                i += 1;
-                continue;
-            }
-            if ch == '"' && !in_single_quote {
-                in_double_quote = !in_double_quote;
-                i += 1;
-                continue;
-            }
-            if in_single_quote || in_double_quote {
-                i += 1;
-                continue;
-            }
-
-            // Found `<<` outside quotes -- check for file-write pattern via
-            // backward scan from the first `<`.
-            if ch == '<' && i + 1 < len && bytes[i + 1] == b'<' {
-                if scan_backward_for_file_write(bytes, i) {
-                    return Err(file_write_heredoc_error());
-                }
-                // Skip past `<<` to avoid re-scanning the same token.
-                // Use a simple increment; the main scan loop below handles
-                // the full delimiter parsing.
-                i += 2;
-                continue;
-            }
-            i += 1;
-        }
-    }
-
-    // Phase 2: original heredoc delimiter scan (quote state is fresh)
-    let mut i = 0;
-    let mut in_single_quote = false;
-    let mut in_double_quote = false;
-
-    while i < len {
-        let ch = bytes[i] as char;
-
-        // Single-quote regions: no escaping inside; toggle on every `'`.
-        if ch == '\'' && !in_double_quote {
-            in_single_quote = !in_single_quote;
-            i += 1;
-            continue;
-        }
-
-        // Double-quote regions: toggle on unescaped `"` outside single quotes.
-        if ch == '"' && !in_single_quote {
-            in_double_quote = !in_double_quote;
-            i += 1;
-            continue;
-        }
-
-        // Inside any quoted region, skip everything (including `<<` tokens).
-        if in_single_quote || in_double_quote {
-            i += 1;
-            continue;
-        }
-
-        // Look for `<<` token
-        if ch == '<' && i + 1 < len && bytes[i + 1] == b'<' {
-            let _here_start = i;
-            i += 2;
-
-            let strip_tabs = if i < len && bytes[i] == b'-' {
-                i += 1;
-                true
-            } else {
-                false
-            };
-
-            // Skip whitespace before delimiter
-            while i < len && (bytes[i] as char).is_ascii_whitespace() {
-                i += 1;
-            }
-
-            if i >= len {
-                return Err(missing_heredoc_error());
-            }
-
-            // Extract the delimiter word, stripping quotes
-            let delimiter = if bytes[i] == b'\'' {
-                // Single-quoted delimiter: <<'EOF'
-                i += 1;
-                let start = i;
-                while i < len && bytes[i] != b'\'' {
-                    i += 1;
-                }
-                if i >= len {
-                    return Err(missing_heredoc_error());
-                }
-                let word = &command[start..i];
-                i += 1; // skip closing quote
-                word.to_string()
-            } else if bytes[i] == b'"' {
-                // Double-quoted delimiter: <<"EOF"
-                i += 1;
-                let start = i;
-                while i < len && bytes[i] != b'"' {
-                    i += 1;
-                }
-                if i >= len {
-                    return Err(missing_heredoc_error());
-                }
-                let word = &command[start..i];
-                i += 1; // skip closing quote
-                word.to_string()
-            } else if bytes[i] == b'\\' {
-                // Escaped delimiter: <<\EOF
-                i += 1;
-                let start = i;
-                while i < len && !(bytes[i] as char).is_ascii_whitespace() && bytes[i] != b'<' {
-                    i += 1;
-                }
-                command[start..i].to_string()
-            } else {
-                // Bare delimiter: <<EOF
-                let start = i;
-                while i < len && !(bytes[i] as char).is_ascii_whitespace() && bytes[i] != b'<' {
-                    i += 1;
-                }
-                command[start..i].to_string()
-            };
-
-            if delimiter.is_empty() {
-                return Err(missing_heredoc_error());
-            }
-
-            // Search the remainder of the command string for a line matching the
-            // closing delimiter.  Walk line-by-line from after the `<<` token,
-            // looking for a line that consists of exactly the delimiter word and
-            // nothing else.
-            //
-            // POSIX rule: the closing delimiter must appear alone on its line
-            // with no leading whitespace (for `<<`) or only leading tabs (for
-            // `<<-`), and no trailing whitespace or comments.  Using a bare
-            // `.trim()` would cause false negatives (accepting `EOF ` or
-            // `  EOF` as closers when the shell does not) so we compare the
-            // stripped line to the delimiter with an exact equality check.
-            //
-            // Using split_inclusive('\n') avoids manual index arithmetic and
-            // eliminates any off-by-one risk on the final line: the iterator
-            // yields every line including the terminating '\n' when present, and
-            // the last segment (no trailing newline) is yielded as-is.
-            let mut found = false;
-            let rest = &command[i..];
-            let mut consumed = i;
-
-            for raw_line in rest.split_inclusive('\n') {
-                // Strip the line terminator only; preserve all other whitespace
-                // so the comparison is exact.
-                let line = raw_line.trim_end_matches('\n');
-                let candidate = if strip_tabs {
-                    // <<-: strip leading tabs only (POSIX; spaces are NOT stripped)
-                    line.trim_start_matches('\t')
-                } else {
-                    line
-                };
-
-                if candidate == delimiter {
-                    found = true;
-                    i = consumed + raw_line.len();
-                    break;
-                }
-
-                consumed += raw_line.len();
-            }
-
-            if !found {
-                return Err(missing_heredoc_error());
-            }
-        } else {
-            i += 1;
-        }
-    }
-
-    Ok(())
-}
-
-fn scan_backward_for_file_write(bytes: &[u8], here_pos: usize) -> bool {
-    /// Checks whether the token before `pos` (walked backward to preceding
-    /// whitespace or command separator) is returned as a byte slice.
-    /// Stops at command separators: whitespace, `(`, `|`, `;`, `&`.
-    fn token_before_pos<'a>(bytes: &'a [u8], pos: &mut usize) -> &'a [u8] {
-        let end = *pos;
-        while *pos > 0 {
-            let b = bytes[*pos - 1];
-            if b.is_ascii_whitespace() || b == b'(' || b == b'|' || b == b';' || b == b'&' {
-                break;
-            }
-            *pos -= 1;
-        }
-        &bytes[*pos..end]
-    }
-
-    /// Extracts a token backward from `pos` that may contain paren-grouped
-    /// constructs such as `$(...)`, `>(...)`, `<(...)`.  When encountering `)`
-    /// the helper enters paren-tracking mode and skips backward through the
-    /// group (including interior whitespace) until the matching `(` is found.
-    /// If `(` is preceded by `$`, `>`, or `<`, the whole construct is consumed
-    /// as part of the token.
-    ///
-    /// NOTE: escaped parentheses (`\(`, `\)`) are not handled.  The heredoc
-    /// security gate operates on raw shell command strings before any
-    /// evaluation, so escape sequences at this level are vanishingly rare in
-    /// practice.  If that assumption ever changes, this function will need a
-    /// backslash-lookahead before decrementing `depth`.
-    fn paren_aware_token<'a>(bytes: &'a [u8], pos: &mut usize) -> &'a [u8] {
-        let end = *pos;
-        let mut depth: i32 = 0;
-
-        while *pos > 0 {
-            let b = bytes[*pos - 1];
-            if b == b')' {
-                depth += 1;
-                *pos -= 1;
-            } else if b == b'(' {
-                depth -= 1;
-                *pos -= 1;
-                if depth == 0 {
-                    // Check if preceded by $, >, or < (nested-context marker)
-                    if *pos > 0 && matches!(bytes[*pos - 1], b'$' | b'>' | b'<') {
-                        *pos -= 1;
-                    }
-                    // Continue scanning backward after reaching depth==0 because
-                    // chained constructs like $(cmd), >(proc) may be followed by
-                    // more backward tokens within the same argument.
-                    continue;
-                }
-            } else if depth > 0 {
-                // Inside parens -- skip interior whitespace
-                *pos -= 1;
-            } else if b.is_ascii_whitespace() || b == b'(' || b == b'|' || b == b';' || b == b'&' {
-                break;
-            } else {
-                *pos -= 1;
-            }
-        }
-        &bytes[*pos..end]
-    }
-
-    /// Skips whitespace scanning backward from `pos`.
-    fn skip_ws_backward(bytes: &[u8], pos: &mut usize) {
-        while *pos > 0 && (bytes[*pos - 1] as char).is_ascii_whitespace() {
-            *pos -= 1;
-        }
-    }
-
-    /// Returns true if `cmd` is a known stdin-consuming command that
-    /// accepts heredoc data and writes to a file specified by `>` or `>>`.
-    ///
-    /// Only commands that read from stdin are listed here because the
-    /// file-write heredoc guard rejects patterns like `cmd > file << EOF`
-    /// where `EOF` gets written to `file` instead of being passed to `cmd`.
-    /// Commands like `cp`, `mv`, and `install` read named file arguments,
-    /// not stdin, so they are excluded -- `cp > file << EOF` does not make
-    /// sense as a heredoc file-write pattern.
-    ///
-    /// Variable-expanded command names (starting with `$`) are included as
-    /// a catch-all for dynamic commands that may consume stdin.
-    fn is_file_write_command(cmd: &[u8]) -> bool {
-        cmd == b"cat"
-            || cmd == b"tee"
-            || cmd == b"printf"
-            || cmd == b"dd"
-            || cmd.first() == Some(&b'$')
-    }
-
-    /// Walks backward from `pos` through argument tokens, calling
-    /// `is_file_write_command` on each.  Returns true if a write command
-    /// is found before hitting a command separator, start of string, or
-    /// an empty token.  Used by both `>>` and `>` redirect branches to
-    /// avoid duplicating the scan logic.
-    fn scan_args_for_write_command(bytes: &[u8], pos: &mut usize) -> bool {
-        loop {
-            let tok = token_before_pos(bytes, pos);
-            if tok.is_empty() {
-                return false;
-            }
-            if is_file_write_command(tok) {
-                return true;
-            }
-            skip_ws_backward(bytes, pos);
-            if *pos == 0 {
-                return false;
-            }
-            let next = bytes[*pos - 1];
-            if next == b'|' || next == b';' || next == b'&' || next == b'(' {
-                return false;
-            }
-        }
-    }
-
-    // here_pos points to the first '<' of '<<'.  Walk backward looking for
-    // a file-write pattern (redirect + >? file + <<).
-
-    let mut pos = here_pos;
-
-    // Skip whitespace between << and the preceding word
-    skip_ws_backward(bytes, &mut pos);
-    if pos == 0 {
-        return false;
-    }
-
-    // Find the file path token immediately before <<.
-    // Use paren_aware_token to handle `$(cmd)`, `>(proc)`, `<(...)` in file position.
-    let file_token = paren_aware_token(bytes, &mut pos);
-    if file_token.is_empty() {
-        return false;
-    }
-
-    // Skip whitespace before the file token
-    skip_ws_backward(bytes, &mut pos);
-    if pos == 0 {
-        return false;
-    }
-
-    // Check for >> or > redirect operator before the file token
-    if pos >= 2 && bytes[pos - 1] == b'>' && bytes[pos - 2] == b'>' {
-        // >> append-redirect operator
-        pos -= 2;
-        skip_ws_backward(bytes, &mut pos);
-        if pos == 0 {
-            // Bare >> file << EOF -- no command before redirect
-            return true;
-        }
-        // Walk backward through all arguments (same logic as > branch below).
-        if scan_args_for_write_command(bytes, &mut pos) {
-            return true;
-        }
-    }
-
-    if bytes[pos - 1] == b'>' {
-        // > write-redirect operator
-        pos -= 1;
-        skip_ws_backward(bytes, &mut pos);
-        if pos == 0 {
-            // Bare > file << EOF -- no command before redirect
-            return true;
-        }
-        // Walk backward through all arguments until we reach a command
-        // separator or the start.  This handles patterns like:
-        //   printf '%s\n' hello > file << EOF
-        // where multiple arguments appear between the command and the redirect.
-        if scan_args_for_write_command(bytes, &mut pos) {
-            return true;
-        }
-    }
-
-    // No redirect operator -- check for tee, dd (dd of=file << EOF), install etc.
-    let cmd = token_before_pos(bytes, &mut pos);
-    if is_file_write_command(cmd) {
-        return true;
-    }
-
-    // Check if the previous token is a flag (starts with '-') for
-    // patterns like `tee -a file << EOF` or `install -m 644 file << EOF`
-    if cmd.len() > 1 && cmd[0] == b'-' {
-        skip_ws_backward(bytes, &mut pos);
-        if pos == 0 {
-            return false;
-        }
-        let prev_cmd = token_before_pos(bytes, &mut pos);
-        return is_file_write_command(prev_cmd);
-    }
-
-    false
-}
-
-fn file_write_heredoc_error() -> ErrorData {
-    ErrorData::new(
-        rmcp::model::ErrorCode::INVALID_PARAMS,
-        "heredoc file-write pattern detected (cat/tee/redirect + <<) -- use edit_overwrite to write files instead of shell heredocs".to_string(),
-        Some(error_meta("validation", false, "use edit_overwrite to write files")),
-    )
-}
-
-fn missing_heredoc_error() -> ErrorData {
-    ErrorData::new(
-        rmcp::model::ErrorCode::INVALID_PARAMS,
-        "heredoc closing delimiter not found -- likely a quoting or escaping issue; use edit_overwrite to write files instead of shell heredocs".to_string(),
-        Some(error_meta("validation", false, "use edit_overwrite to write files")),
-    )
-}
-
 /// Validates that the parent directory of `path` exists, is a directory,
 /// and is within `root`.  Returns the resolved path (canonical_parent.join(file_name)).
 fn validate_parent_in_root(
@@ -704,6 +270,7 @@ pub(crate) fn validate_path_in_dir(
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::shell_write::scan_backward_for_file_write;
 
     #[test]
     fn validate_path_no_trailing_slash() {
